refactor: (#743) conflict 해결
