+++ backend/src/main/java/com/votogether/domain/alarm/controller/AlarmQueryController.java
@@ -1,5 +1,7 @@
 package com.votogether.domain.alarm.controller;
 
+import com.votogether.domain.alarm.dto.ReportActionAlarmResponse;
+import com.votogether.domain.alarm.dto.ReportActionResponse;
 import com.votogether.domain.alarm.dto.response.PostAlarmResponse;
 import com.votogether.domain.alarm.service.AlarmQueryService;
 import com.votogether.domain.member.entity.Member;
@@ -10,6 +12,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -27,8 +30,26 @@ public class AlarmQueryController implements AlarmQueryControllerDocs {
             @RequestParam @PositiveOrZero(message = "페이지는 0이상 정수만 가능합니다.") final int page,
             @Auth final Member loginMember
     ) {
-        final List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(page, loginMember);
+        final List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(loginMember, page);
         return ResponseEntity.ok(postAlarmResponses);
+    }
+
+    @GetMapping("/report")
+    public ResponseEntity<List<ReportActionAlarmResponse>> getReportActionAlarms(
+            @RequestParam @PositiveOrZero(message = "페이지는 0이상 정수만 가능합니다.") final int page,
+            @Auth final Member loginMember
+    ) {
+        final List<ReportActionAlarmResponse> response = alarmQueryService.getReportActionAlarms(loginMember, page);
+        return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/report/{id}")
+    public ResponseEntity<ReportActionResponse> getReportActionAlarm(
+            @PathVariable("id") final Long reportActionAlarmId,
+            @Auth final Member loginMember
+    ) {
+        final ReportActionResponse response = alarmQueryService.getReportActionAlarm(reportActionAlarmId, loginMember);
+        return ResponseEntity.ok(response);
     }
 
 }
