+++ backend/src/test/java/com/votogether/domain/alarm/service/AlarmQueryServiceTest.java
@@ -3,16 +3,21 @@ package com.votogether.domain.alarm.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
+import com.votogether.domain.alarm.dto.ReportActionAlarmResponse;
+import com.votogether.domain.alarm.dto.ReportActionResponse;
 import com.votogether.domain.alarm.dto.response.PostAlarmDetailResponse;
 import com.votogether.domain.alarm.dto.response.PostAlarmResponse;
+import com.votogether.domain.alarm.entity.ReportActionAlarm;
 import com.votogether.domain.alarm.repository.ReportActionAlarmRepository;
 import com.votogether.domain.member.entity.Member;
 import com.votogether.domain.post.dto.request.comment.CommentCreateRequest;
 import com.votogether.domain.post.entity.Post;
 import com.votogether.domain.post.service.PostCommandService;
 import com.votogether.domain.post.service.PostCommentService;
+import com.votogether.domain.report.entity.vo.ReportType;
 import com.votogether.test.ServiceTest;
 import java.util.List;
+import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -54,7 +59,7 @@ class AlarmQueryServiceTest extends ServiceTest {
             TestTransaction.end();
 
             // when
-            List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(0, commentWriter);
+            List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(commentWriter, 0);
 
             // then
             PostAlarmResponse postAlarmResponse = postAlarmResponses.get(0);
@@ -81,7 +86,7 @@ class AlarmQueryServiceTest extends ServiceTest {
             Thread.sleep(1000);
 
             // when
-            List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(0, member);
+            List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(member, 0);
 
             // then
             PostAlarmResponse postAlarmResponse = postAlarmResponses.get(0);
@@ -100,12 +105,72 @@ class AlarmQueryServiceTest extends ServiceTest {
             Member member = memberTestPersister.builder().save();
 
             // when
-            List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(0, member);
+            List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(member, 0);
 
             // then
             assertThat(postAlarmResponses).hasSize(0);
         }
 
+    }
+
+    @Test
+    @DisplayName("신고 조치 알림 목록을 조회한다.")
+    void getReportActionAlarms() {
+        // given
+        Member member = memberTestPersister.builder().save();
+
+        ReportActionAlarm reportActionAlarmA = ReportActionAlarm.builder()
+                .reportType(ReportType.POST)
+                .member(member)
+                .isChecked(false)
+                .reasons("광고성, 부적합성")
+                .target("1")
+                .build();
+
+        reportActionAlarmRepository.save(reportActionAlarmA);
+
+        // when
+        List<ReportActionAlarmResponse> reportActionAlarms = alarmQueryService.getReportActionAlarms(member, 0);
+
+        // then
+        ReportActionAlarmResponse result = reportActionAlarms.get(0);
+        assertSoftly(softly -> {
+            softly.assertThat(reportActionAlarms.size()).isEqualTo(1);
+            softly.assertThat(result.alarmId()).isNotNull();
+            softly.assertThat(result.isChecked()).isFalse();
+            softly.assertThat(result.detail().reportActionId()).isNotNull();
+            softly.assertThat(result.detail().content()).isEqualTo("1");
+            softly.assertThat(result.detail().reasons()).containsAll(Set.of("광고성", "부적합성"));
+            softly.assertThat(result.detail().type()).isEqualTo(ReportType.POST);
+        });
+    }
+
+    @Test
+    @DisplayName("신고 조치 알림을 상세 조회한다.")
+    void getReportActionAlarm() {
+        // given
+        Member member = memberTestPersister.builder().save();
+
+        ReportActionAlarm reportActionAlarmA = ReportActionAlarm.builder()
+                .reportType(ReportType.POST)
+                .member(member)
+                .isChecked(false)
+                .reasons("광고성, 부적합성")
+                .target("1")
+                .build();
+
+        ReportActionAlarm savedReportActionAlarm = reportActionAlarmRepository.save(reportActionAlarmA);
+
+        // when
+        ReportActionResponse response = alarmQueryService.getReportActionAlarm(savedReportActionAlarm.getId(), member);
+
+        // then
+        assertSoftly(softly -> {
+            softly.assertThat(response.reportActionId()).isEqualTo(savedReportActionAlarm.getId());
+            softly.assertThat(response.type()).isEqualTo(ReportType.POST);
+            softly.assertThat(response.content()).isEqualTo("1");
+            softly.assertThat(response.reasons()).containsAll(Set.of("광고성", "부적합성"));
+        });
     }
 
 }
